+++ src/main/java/insane96mcp/iguanatweaksreborn/module/farming/livestock/LivestockData.java
@@ -29,8 +29,9 @@ public class LivestockData {
 	final List<Modifier> cowFluidCooldownModifiers;
 	final List<Modifier> sheepWoolGrowthChanceModifiers;
 	final List<Modifier> twinsChanceModifiers;
+	final List<Modifier> autoBreedChanceModifiers;
 
-    public LivestockData(IdTagMatcher entity, @Nullable Float livingDays, float livingDaysFluctuation, List<Modifier> growthSpeedModifiers, List<Modifier> breedingCooldownModifiers, List<Modifier> eggLayCooldownModifiers, List<Modifier> breedingFailChanceModifiers, List<Modifier> cowFluidCooldownModifiers, List<Modifier> sheepWoolGrowthChanceModifiers, List<Modifier> twinsChanceModifiers) {
+    public LivestockData(IdTagMatcher entity, @Nullable Float livingDays, float livingDaysFluctuation, List<Modifier> growthSpeedModifiers, List<Modifier> breedingCooldownModifiers, List<Modifier> eggLayCooldownModifiers, List<Modifier> breedingFailChanceModifiers, List<Modifier> cowFluidCooldownModifiers, List<Modifier> sheepWoolGrowthChanceModifiers, List<Modifier> twinsChanceModifiers, List<Modifier> autoBreedChanceModifiers) {
         this.entity = entity;
 		this.livingDays = livingDays;
 		this.livingDaysFluctuation = livingDaysFluctuation;
@@ -41,6 +42,7 @@ public class LivestockData {
         this.cowFluidCooldownModifiers = cowFluidCooldownModifiers;
         this.sheepWoolGrowthChanceModifiers = sheepWoolGrowthChanceModifiers;
         this.twinsChanceModifiers = twinsChanceModifiers;
+		this.autoBreedChanceModifiers = autoBreedChanceModifiers;
     }
 
 	public boolean matches(Entity entity) {
@@ -71,7 +73,8 @@ public class LivestockData {
 			List<Modifier> cowFluidCooldownModifiers = Modifier.deserializeList(jObject, "cow_fluid_cooldown_modifiers", context);
 			List<Modifier> sheepWoolGrowthChanceModifiers = Modifier.deserializeList(jObject, "sheep_wool_growth_chance_modifiers", context);
 			List<Modifier> twinsChanceModifiers = Modifier.deserializeList(jObject, "twins_chance_modifiers", context);
-			return new LivestockData(entity, livingDays, livingDaysFluctuation, growthSpeed, breedingCooldown, eggLayCooldown, breedingFailChanceModifiers, cowFluidCooldownModifiers, sheepWoolGrowthChanceModifiers, twinsChanceModifiers);
+			List<Modifier> autoBreedChanceModifiers = Modifier.deserializeList(jObject, "auto_breed_chance_modifiers", context);
+			return new LivestockData(entity, livingDays, livingDaysFluctuation, growthSpeed, breedingCooldown, eggLayCooldown, breedingFailChanceModifiers, cowFluidCooldownModifiers, sheepWoolGrowthChanceModifiers, twinsChanceModifiers, autoBreedChanceModifiers);
 		}
 	}
 }